基本完成代理授权的UI绘制
1、基本完成代理授权的UI绘制
2、还有参数传递和接口需要增加

+++ WeiShang/Controllers/Products/ProductsListViewController.m
@@ -299,7 +299,7 @@
 -(void)onClickUser:(NSUInteger)userId
 {
     //点击左边头像 或者 点击评论和赞的用户昵称
-    NSLog(@"onClickUser: %ld", userId);
+    NSLog(@"onClickUser: %lu", (unsigned long)userId);
     
     UserViewController *controller = [[UserViewController alloc] init];
     [self.navigationController pushViewController:controller animated:YES];
